Update PCB
Done Majority of routing on top layer of PCB. Changed the UART pins we were using to 55,56,58,59 (also added flow control with 0 ohm jumpers. Still have to do a board cutout for the antenna on the ESP-wroom-O2D. And put a header on the ESP reset pin in case we need it.