Finishing touches on PCB
Adding a few jumpers to the power supply so we can measure current for qualifiers. Also not sure what to do with the RST pin on the ESP so im hooking them up to header pins just incase we want to use SPI for communication. Will finish with a polygon pour.