Fixed App Issues
MyFirstApp would not connect to firebase anymore and I think it was due to the fact that I was running the emulator at API level 16. When I switched to API level 29 I was able to connect to firebase. Also was able to set up an event listener to get real-time updates. Next add event listeners for other properties and add some UI elements.